Features and fixes
Added new feature: Initial sort order for memberlist (issue #29 of the
original repository)
Fixed Incorrect detail links on "Disable page alias usage" (issue #27,
#28 of the original repository, fix by BugBuster)
Added template variable total_members to offer the total number of
member datasets for the mod_memberlist template (issue #29 of the
original repository)
Added sender email address for member details contact form (issue #18
of the original repository)
Added new Hook memberlistQuery to customize the output of the
memberlist via other modules. The Hook function uses 3 parameters
memberlistQuery(ModuleMemberlist $memberlistObject, string $where,
array $where_parameters). The first parameter is a reference to the
memberlist module instance, the second parameter is the WHERE SQL
expression and should be used as reference to be changed, the third
parameter is the array containing the parameters for the where
expression and should be used as reference to be changed.